+++ dart/math/Helpers.h
@@ -183,12 +183,12 @@ inline bool isSymmetric(const Eigen::MatrixXd& _m, double _tol = 1e-6) {
   if (rows != cols)
     return false;
 
-  for (int i = 0; i < rows; ++i) {
-    for (int j = i + 1; j < cols; ++j) {
+  for (size_t i = 0; i < rows; ++i) {
+    for (size_t j = i + 1; j < cols; ++j) {
       if (std::fabs(_m(i, j) - _m(j, i)) > _tol) {
         std::cout << "A: " << std::endl;
-        for (int k = 0; k < rows; ++k) {
-          for (int l = 0; l < cols; ++l)
+        for (size_t k = 0; k < rows; ++k) {
+          for (size_t l = 0; l < cols; ++l)
             std::cout << std::setprecision(4) << _m(k, l) << " ";
           std::cout << std::endl;
         }
